Don't render the children on initial render.
When the Frame component is rendered, do not include this.props.children on the iframe's props. The DOM of the iframe is not ready at that point, so the children will be rendered as detached from the DOM tree.

This will trigger issues such as "Invariant Violation: findComponentRoot(..., .0.0): Unable to find element" when React.findDOMNode(this) is called inside componentDidMount().

The children should only be rendered when the the iframe is ready, and within Frame.renderFrameContents.