Add support for setting initial content of the frame,  fixes #15
This introduces a new prop "initialContent". initialContent is injected into the
frame the first time it is rendered. That allows you to insert script tags or
whatever else into the head element. This does not attempt to update the frame
when the initialContent prop changes (hence the prop starts with "initial").