Don't unmountComponentAtNode unless the mount target exists
In some automated tests I was running I found that every once in a while
this would happen:
1. frame component is mounted
2. iframe / doc are created (but subtree is not yet rendered into iframe)
3. frame component is unmounted

In this case unmounting the frame component would lead to an error since
the mount target would be undefined.

This commit checks that both the doc and mountTarget exist before
calling unmountComponentAtNode.