Fix TypeScript Error
Hi, I have TypeScript 4.6.4, and I get this error:
`'Frame' cannot be used as a JSX component. Its instance type 'FrameComponent' is not a valid JSX element.`

I fixed this with these changes. I'm not a TypeScript expert, I'm not sure if it breaks types for older versions or not.